PeriodicChecker implementation
PeriodicChecker is a go routine that checks every 3 seconds for
incosistencies between the memory port mappings (ContainerPorts) and the
ones used by Docker. It syncs the ContainerPorts and then it checks for
expired sessions in Redis. If a session has expired, it will kill the
corresponding containers. Then the ContainerPorts are again in sync.

    * moved session package into dc
    * implemented PeriodicChecker go routine